requirements.txt
click==7.1.1
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Uploads==0.2.1
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
Pillow==7.1.1
python-dotenv==0.13.0
visitor==0.1.3
Werkzeug==0.15.6
torch==1.5.0
torchvision==0.6.0
